+++ src/editor/markdown/index.js
@@ -2,13 +2,10 @@
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
- *
  */
 
-export {getParser} from "./parser"
-export {getSerializer} from "./serializer"
-export {getRenderer} from "./renderer"
-export {createLinkExtension} from "./linkExtensionTokenizer"
-export * from "prosemirror-markdown"
-
-
+export {getParser} from "./parser";
+export {getSerializer} from "./serializer";
+export {getRenderer} from "./renderer";
+export {createLinkExtension} from "./linkExtensionTokenizer";
+export * from "prosemirror-markdown";