+++ src/editor/markdown/renderer.js
@@ -2,30 +2,29 @@
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
- *
  */
 
-import markdownit from "markdown-it"
-import {getPlugins, PresetManager} from "../core/plugins"
+import markdownit from "markdown-it";
+import {getPlugins, PresetManager} from "../core/plugins";
 
-let presets = new PresetManager({
+const presets = new PresetManager({
     name: 'renderer',
-    create: (options) => {
-        return createRenderer(options);
+    create: (context) => {
+        return createRenderer(context);
     }
 });
 
-let getRenderer = (context) => {
+const getRenderer = (context) => {
     return presets.check(context);
 };
 
-let createRenderer = function(context) {
-    let markdownItOptions = context && context.options.markdownIt || {html: false, breaks: true, linkify: true};
-    let renderer = markdownit(markdownItOptions);
+const createRenderer = (context) => {
+    const markdownItOptions = context && context.options.markdownIt || {html: false, breaks: true, linkify: true};
+    const renderer = markdownit(markdownItOptions);
 
     const plugins = getPlugins(context);
     plugins.forEach((plugin) => {
-        if(plugin.registerMarkdownIt) {
+        if (plugin.registerMarkdownIt) {
             plugin.registerMarkdownIt(renderer);
         }
     });
@@ -33,4 +32,4 @@ let createRenderer = function(context) {
     return renderer;
 };
 
-export {getRenderer}
+export {getRenderer};