+++ docs/CHANGELOG.md
@@ -28,6 +28,7 @@ Changelog
 - Enh #43: Make menu accessible
 - Fix #51: Link creation does not copy URL from selected text
 - Enh #47: Edit code block in codemirror
+- Fix #113: Header anchors lost
 
 1.1.7 (August 31, 2021)
 -----------------------